--- Assets/Scripts/JoyCon/JoyConReader.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 9b480349498c74da7b99d9fea837870a
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
